(PC-17471)[PRO] chore: move CollectiveOfferSummary in new_components and remove non shared elements
