+++ pro/src/new_components/CollectiveOfferSummary/CollectiveOfferSummary.tsx
@@ -1,22 +1,14 @@
 import React from 'react'
 
 import {
-  cancelCollectiveBookingAdapter,
   CollectiveOffer,
   CollectiveOfferTemplate,
   EducationalCategories,
-  patchIsCollectiveOfferActiveAdapter,
-  patchIsTemplateOfferActiveAdapter,
 } from 'core/OfferEducational'
-import { computeURLCollectiveOfferId } from 'core/OfferEducational/utils/computeURLCollectiveOfferId'
-import useActiveFeature from 'hooks/useActiveFeature'
-import useNotification from 'hooks/useNotification'
-import OfferEducationalActions from 'new_components/OfferEducationalActions'
 import { SummaryLayout } from 'new_components/SummaryLayout'
 import { ButtonLink } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
 
-import styles from './CollectiveOfferSummary.module.scss'
 import CollectiveOfferAccessibilitySection from './components/CollectiveOfferAccessibilitySection'
 import CollectiveOfferContactSection from './components/CollectiveOfferContactSection'
 import CollectiveOfferNotificationSection from './components/CollectiveOfferNotificationSection'
@@ -31,101 +23,20 @@ import { DEFAULT_RECAP_VALUE } from './components/constants'
 interface ICollectiveOfferSummaryProps {
   offer: CollectiveOfferTemplate | CollectiveOffer
   categories: EducationalCategories
-  reloadCollectiveOffer?: () => void
+  offerEditLink?: string
+  stockEditLink?: string
+  visibilityEditLink?: string
 }
 
 const CollectiveOfferSummary = ({
   offer,
   categories,
-  reloadCollectiveOffer,
+  offerEditLink,
+  stockEditLink,
+  visibilityEditLink,
 }: ICollectiveOfferSummaryProps) => {
-  const notify = useNotification()
-
-  /* istanbul ignore next: DEBT, TO FIX */
-  const setIsOfferActive = async () => {
-    const adapter = offer.isTemplate
-      ? patchIsTemplateOfferActiveAdapter
-      : patchIsCollectiveOfferActiveAdapter
-
-    const response = await adapter({
-      offerId: offer.id,
-      isActive: !offer.isActive,
-    })
-
-    if (response.isOk) {
-      return notify.success(response.message)
-    }
-
-    notify.error(response.message)
-  }
-
-  /* istanbul ignore next: DEBT, TO FIX */
-  const cancelActiveBookings = async () => {
-    if (offer.isTemplate) {
-      return
-    }
-
-    const { isOk, message } = await cancelCollectiveBookingAdapter({
-      offerId: offer.id,
-    })
-
-    if (!isOk) {
-      return notify.error(message)
-    }
-
-    notify.success(message)
-    reloadCollectiveOffer?.()
-  }
-
-  const isCollectiveOfferDuplicationActive = useActiveFeature(
-    'WIP_CREATE_COLLECTIVE_OFFER_FROM_TEMPLATE'
-  )
-
-  const offerEditLink = `/offre/${computeURLCollectiveOfferId(
-    offer.id,
-    offer.isTemplate
-  )}/collectif/edition`
-
-  const stockEditLink = `/offre/${computeURLCollectiveOfferId(
-    offer.id,
-    offer.isTemplate
-  )}/collectif/stocks/edition`
-
-  const visibilityEditLink = `/offre/${offer.id}/collectif/visibilite/edition`
-
   return (
     <>
-      <OfferEducationalActions
-        cancelActiveBookings={cancelActiveBookings}
-        className={styles.actions}
-        isBooked={
-          offer.isTemplate ? false : Boolean(offer.collectiveStock?.isBooked)
-        }
-        isCancellable={offer.isCancellable}
-        isOfferActive={offer.isActive}
-        setIsOfferActive={setIsOfferActive}
-      />
-      {isCollectiveOfferDuplicationActive && offer.isTemplate && (
-        <div className={styles['duplicate-offer']}>
-          <p className={styles['duplicate-offer-description']}>
-            Vous pouvez dupliquer cette offre autant de fois que vous le
-            souhaitez pour l’associer aux établissements scolaires qui vous
-            contactent. <br />
-            &nbsp;· L’offre vitrine restera visible sur ADAGE <br />
-            &nbsp;· L’offre associée à l’établissement devra être pré-réservée
-            par l’enseignant(e) qui vous a contacté
-          </p>
-          <ButtonLink
-            variant={ButtonVariant.PRIMARY}
-            link={{
-              isExternal: false,
-              to: `/offre/duplication/collectif/${offer.id}`,
-            }}
-          >
-            Créer une offre réservable pour un établissement scolaire
-          </ButtonLink>
-        </div>
-      )}
       <SummaryLayout>
         <SummaryLayout.Content fullWidth>
           <SummaryLayout.Section
@@ -174,12 +85,6 @@ const CollectiveOfferSummary = ({
           )}
         </SummaryLayout.Content>
       </SummaryLayout>
-      <ButtonLink
-        variant={ButtonVariant.PRIMARY}
-        link={{ isExternal: false, to: '/offers/collective' }}
-      >
-        Retour à la liste des offres
-      </ButtonLink>
     </>
   )
 }
