+++ pro/src/new_components/CollectiveOfferSummary/__specs__/CollectiveOfferSummary.spec.tsx
