+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferAccessibilitySection/CollectiveOfferAccessibilitySection.tsx
