+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferAccessibilitySection/index.ts
