+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferContactSection/index.ts
