+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferNotificationSection/CollectiveOfferNotificationSection.tsx
