+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferNotificationSection/index.ts
