+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferParticipantSection/CollectiveOfferParticipantSection.tsx
