+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferParticipantSection/index.ts
