+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferPracticalInformation/CollectiveOfferPracticalInformation.tsx
