+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferPracticalInformation/index.ts
