+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferStockSection/CollectiveOfferStockSection.tsx
