+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferStockSection/__specs__/CollectiveOfferStockSection.spec.tsx
