+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferStockSection/index.ts
