+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferTypeSection/CollectiveOfferTypeSection.tsx
