+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferTypeSection/index.ts
