+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferVenueSection/CollectiveOfferVenueSection.tsx
