+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferVenueSection/index.ts
