+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferVisibilitySection/CollectiveOfferVisibilitySection.tsx
