+++ pro/src/new_components/CollectiveOfferSummary/components/CollectiveOfferVisibilitySection/index.ts
