+++ pro/src/new_components/CollectiveOfferSummary/components/constants.ts
