+++ pro/src/new_components/CollectiveOfferSummary/components/utils/__specs__/formatDuration.spec.tsx
