+++ pro/src/new_components/CollectiveOfferSummary/components/utils/__specs__/formatOfferEventAddress.spec.ts
