+++ pro/src/new_components/CollectiveOfferSummary/components/utils/formatDuration.ts
