+++ pro/src/new_components/CollectiveOfferSummary/components/utils/formatOfferEventAddress.ts
