+++ pro/src/new_components/CollectiveOfferSummary/index.ts
