+++ pro/src/routes/CollectiveOfferSummaryEdition/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CollectiveOfferSummaryEdition'
