--- pro/src/routes/CollectiveOfferTemplateSummary/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveOfferTemplateSummary'
