--- pro/src/screens/CollectiveOfferSummary/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveOfferSummary'
